requirements.txt
aiosmtplib==5.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==3.2.2
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-mail==1.6.1
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
passlib==1.7.4
pgvector==0.4.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
razorpay==2.0.0
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
